Generate Products
The products table will have two two fields. The name and stripe id. The idea is that you setup
the products in Stripe and then just store a reference to it in your local database. To generate
the products inside the billing context, i run the command:
mix phx.gen.context Billing Product products --table billing_products \
stripe_id stripe_product_name

Note that I specify table name to be billing_products . I do this because I might have
another table called products in the future and that might not be assiciated with the Stripe
checkout.